+++ test/cases/stream/stream.go
@@ -50,8 +50,9 @@ var streamEntries = []any{
 	g.Entry("excludes data expired beyond TTL", helpers.Args{Input: "all", Offset: -156 * time.Hour, Duration: 24 * time.Hour, WantEmpty: true}),
 	g.Entry("projection with http.method", helpers.Args{Input: "all_with_http_method", Duration: 1 * time.Hour}),
 	g.Entry("limit", helpers.Args{Input: "limit", Duration: 1 * time.Hour}),
-	g.Entry("max limit", helpers.Args{
-		Input: "all_max_limit", WantErr: true, WantErrCode: codes.InvalidArgument, Duration: 1 * time.Hour,
+	g.Entry("max limit", helpers.Args{Input: "all_max_limit", Want: "all", Duration: 1 * time.Hour}),
+	g.Entry("above absolute window", helpers.Args{
+		Input: "all_above_window", WantErr: true, WantErrCode: codes.InvalidArgument, Duration: 1 * time.Hour,
 	}),
 	g.Entry("offset", helpers.Args{Input: "offset", Duration: 1 * time.Hour}),
 	g.Entry("order asc", helpers.Args{Input: "order_asc", Duration: 1 * time.Hour}),